Created method classifier_classify() in models/classifier.py
This method is a driver function to classify the data.

It calls the methods defined above it in the class. It then stores 2 data members of the object in 2 variables:
accuracy, prediction = self.accuracy, self.prediction
and returns them  to a calling function.

This function is supposed to help the user quickly classify the input data, without having to acquire knowledge of the underlying class functions.