Moved models/ under server/ folder
Python is unable to import packages placed in siblings of parent directory. Thus, server.py (in root/server/server.py) is unable to import models.classifier(root/models/classifier.py). So we moded the models/ directory under server/ folder.

There are many ways to inject the sibling of a directory in the PATH so that python can detect it. However, since our goal isn't to explore this side of programming, we have chosen to make a simple change that allows us to run our app with minimal change to environment variables and requires no additional steps beyond simple existence.